changes to java sdk to compile against matchers
