+++ sdks/java/client-java-loadtest/src/main/java/io/featurehub/loadtest/LoadTest.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import io.featurehub.client.AbstractFeatureRepository;
 import io.featurehub.client.AnalyticsCollector;
 import io.featurehub.client.ClientContext;
 import io.featurehub.client.ClientContextRepository;
@@ -54,7 +55,7 @@ class FeatureVersionHolder {
   }
 }
 
-class InternetFeatureTrackerRepository implements FeatureRepository {
+class InternetFeatureTrackerRepository extends AbstractFeatureRepository {
   private static final Logger log = LoggerFactory.getLogger(InternetFeatureTrackerRepository.class);
   public static int fullDataCount = 0;
   Map<Long, FeatureVersionHolder> versionMap = new ConcurrentHashMap<>();
